# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InstaClone API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor de API Principal"
        }
    ],
    "paths": {
...
        "/users/{id}/follow": {
            "post": {
                "tags": [
                    "Follow"
                ],
                "summary": "Seguir usuário",
                "description": "Permite ao usuário autenticado seguir outro usuário. Requer autenticação.",
                "operationId": "6f109447b27b2dcc26dbf7dfeb0ce585",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário a ser seguido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário seguido com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Seguindo com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "422": {
                        "description": "Erro ao seguir (pode estar seguindo já)"
                    }
                }
            }
        },
        "/users/{id}/unfollow": {
            "delete": {
                "tags": [
                    "Follow"
                ],
                "summary": "Deixar de seguir usuário",
                "description": "Remove o usuário autenticado da lista de seguidores de outro usuário. Requer autenticação.",
                "operationId": "a370b6512b85f3dccf6bb694dc770a43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário para deixar de seguir",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parou de seguir com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Deixou de seguir"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
        "/users/{id}/is-following": {
            "get": {
                "tags": [
                    "Follow"
                ],
                "summary": "Verificar status de seguimento",
                "description": "Verifica se o usuário autenticado está seguindo um usuário específico. Requer autenticação.",
                "operationId": "5c362493b3117c7985d78ffab45cca9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário a verificar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status de seguimento retornado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_following": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
        "/users/{id}/followers": {
            "get": {
                "tags": [
                    "Follow"
                ],
                "summary": "Listar seguidores do usuário",
                "description": "Retorna a lista paginada de usuários que seguem o usuário especificado.",
                "operationId": "6e630eb56369e97c29b81c98e068fbd1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (padrão: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de seguidores retornada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Maria Silva"
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "maria_silva"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "maria@example.com"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "/storage/avatars/maria.jpg"
                                                    },
                                                    "bio": {
                                                        "type": "string",
                                                        "example": "Fotógrafa e viajante"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-05-22T10:30:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 42
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
        "/users/{id}/following": {
            "get": {
                "tags": [
                    "Follow"
                ],
                "summary": "Listar usuários seguidos",
                "description": "Retorna a lista paginada de usuários que o usuário especificado está seguindo.",
                "operationId": "2b478d3762ffcd6828796157b9f2529b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página (padrão: 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários seguidos retornada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "João Pedro"
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "joao_pedro"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "joao@example.com"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "/storage/avatars/joao.jpg"
                                                    },
                                                    "bio": {
                                                        "type": "string",
                                                        "example": "Desenvolvedor Full Stack"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-05-20T15:45:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 28
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}